[feat] Implementing ExceptionVerifier

+++ ApiFirst/AppServices/Interfaces/ICustomerAppService.cs
@@ -1,15 +1,16 @@
 using AppModels.AppModels;
+using System.Collections.Generic;
 
 namespace AppServices.Interfaces
 {
     public interface ICustomerAppService
     {
         long Create(CreateCustomerDTO model);
-        void Update(int id, UpdateCustomerDTO model);
-        void Delete(int id);
+        void Update(long id, UpdateCustomerDTO model);
+        void Delete(long id);
         IEnumerable<CustomerResponseDTO> GetAll();
-        CustomerResponseDTO GetById(int id);
-        void UpdateEmail(int id, string email);
+        CustomerResponseDTO GetById(long id);
+        void UpdateEmail(long id, string email);
         CustomerResponseDTO GetByCpf(string cpf);
     }
 }
